Use default settings for interactions and tables list dialog
Fixes #6387
