+++ specifyweb/frontend/js_src/lib/localization/forms.ts
@@ -1085,33 +1085,6 @@ export const formsText = createDictionary({
     'fr-fr': 'Sous-vue',
     'pt-br': 'Subvisualização',
   },
-  selectSourceOfTables: {
-    'en-us': 'Select source of tables',
-    'ru-ru': 'Выберите источник таблиц',
-    'es-es': 'Seleccionar fuente de tablas',
-    'fr-fr': 'Sélectionner la source des tables',
-    'uk-ua': 'Виберіть джерело таблиць',
-    'de-ch': 'Wählen Sie die Quelle der Tabellen aus',
-    'pt-br': 'Selecione a fonte das tabelas',
-  },
-  inheritLegacySettings: {
-    'en-us': 'Copy Specify 6 settings',
-    'ru-ru': 'Копировать Укажите 6 настроек',
-    'es-es': 'Copiar Especificar 6 configuraciones',
-    'fr-fr': 'Copier Spécifier 6 paramètres',
-    'uk-ua': 'Копіювати Вказати 6 налаштувань',
-    'de-ch': 'Kopieren Geben Sie 6 Einstellungen an',
-    'pt-br': 'Copiar Especificar 6 configurações',
-  },
-  useCustomSettings: {
-    'en-us': 'Use custom settings',
-    'ru-ru': 'Использовать пользовательские настройки',
-    'es-es': 'Usar configuraciones personalizadas',
-    'fr-fr': 'Utiliser les paramètres personnalisés',
-    'uk-ua': 'Використовуйте власні налаштування',
-    'de-ch': 'Benutzerdefinierte Einstellungen verwenden',
-    'pt-br': 'Usar configurações personalizadas',
-  },
   disableReadOnly: {
     'en-us': 'Disable read-only mode',
     'ru-ru': 'Отключить режим только для чтения',
@@ -1129,24 +1102,6 @@ export const formsText = createDictionary({
     'uk-ua': 'Увімкнути режим лише для читання',
     'de-ch': 'Aktivieren Sie den Nur-Lese-Modus',
     'pt-br': 'Habilitar modo somente leitura',
-  },
-  configureDataEntryTables: {
-    'en-us': 'Configure data entry tables',
-    'ru-ru': 'Настройте таблицы ввода данных',
-    'es-es': 'Configurar tablas de entrada de datos',
-    'fr-fr': 'Configurer les tables de saisie de données',
-    'uk-ua': 'Налаштування таблиць для введення даних',
-    'de-ch': 'Konfigurieren von Dateneingabetabellen',
-    'pt-br': 'Configurar tabelas de entrada de dados',
-  },
-  configureInteractionTables: {
-    'en-us': 'Configure interaction tables',
-    'ru-ru': 'Настройте таблицы взаимодействия',
-    'es-es': 'Configurar tablas de interacción',
-    'fr-fr': "Configurer les tables d'interaction",
-    'uk-ua': 'Налаштування таблиць взаємодії',
-    'de-ch': 'Konfigurieren von Interaktionstabellen',
-    'pt-br': 'Configurar tabelas de interação',
   },
   formMeta: {
     'en-us': 'Form Meta',
